Began revised AST
We began rewriting the AST and parser to flatten the whole program
representation into just a bunch of lookup tables keyed with lightweight
UUIDs. That will make a couple of things easier.

* First, I can entirely stop worrying about ownership issues between
  various syntax trees and various symbol table things, basically because
  everything is allocated from my own "arenas," I'm the one issuing the
  pointers, and I can make them Copy if I want to, gosh darn it.
* Second, it'll be infinitely easier to do things like "return a
  function," or "move a value of a local type outside the scope where the
  type is defined," because there's just one global table of "function
  pointers" (resp. ADTs). No need to sprinkle Rc<RefCell<_>>s everywhere.

Now. How and when to do that? It could be done during a "lowering" phase
from the AST I have to some high-level, lightly-desugared IR. I actually
tried writing that out, and I don't like it. The main reason is that
this IR is *almost identical* to the AST I started from, but not similar
enough that both can be abstracted by generics or something. So keeping
them in sync is just *more work to do*. It's also one more tree
traversal, which means slower compiles. An alternative would be to lower
the highest-level representation! In other words, never build an
AST, *exactly*. Let the first internal representation of the program be
a forest, and don't worry about it. The result will be something more
like rustc's HIR than its AST. This might be a little unorthodox, but
I'm going to try it.

What this will look like is that all parser instances (there may be
several!) take a mutable reference to some `AstCtx` struct. The
beginning of the refactoring process has had several effects throughout
the codebase:
* As a first go at using the `AstCtx`, identifiers are now interned.
* Many of the AST structs have been replaced with a generic `Spanned<T>` type.
* A lot of modules have been disabled entirely. There is a tradeoff
  between development speed (get rid of the compiler errors in the middle
  of refactoring) and safety. We'll see how it goes. Maybe I'll learn
  from it!

+++ cavy/src/store.rs
@@ -11,7 +11,7 @@ use std::marker::PhantomData;
 ///
 /// As I somehow keep converging on the same design choices as rustc, it might
 /// be worth pointing out that this is exactly analogous to `rustc_index::Idx`.
-pub trait Index<V>: From<usize> + Eq + Hash + Copy {}
+pub trait Index<V>: From<u32> + Eq + Hash + Copy {}
 
 /// A macro for building implementers of Index. This is exactly analogous to
 /// rustc's `rustc_index::newtype_index`.
@@ -19,10 +19,10 @@ pub trait Index<V>: From<usize> + Eq + Hash + Copy {}
 macro_rules! store_triple {
     ($store:ident : $index:ident => $V:ty) => {
         #[derive(PartialEq, Eq, Hash, Clone, Copy, Debug, Default)]
-        pub struct $index(usize);
+        pub struct $index(u32);
 
-        impl From<usize> for $index {
-            fn from(val: usize) -> Self {
+        impl From<u32> for $index {
+            fn from(val: u32) -> Self {
                 Self(val)
             }
         }
@@ -43,7 +43,7 @@ pub struct Store<Idx, V>
 where
     Idx: Index<V>,
 {
-    next_internal: usize,
+    next_internal: u32,
     backing_store: HashMap<Idx, V>,
     phantom: PhantomData<Idx>,
 }
